Add popover to mobile bounds layer
